fix(seo): emit JSON-LD on currency-hedging-cost calculator
The calculator shipped as a single "use client" page with no metadata or
JSON-LD, tripping the JSON-LD coverage gate. Split into a server page that
exports metadata + SoftwareApplication/breadcrumb/FAQ JSON-LD (matching the
other calculator pages) rendering the interactive client component.
